Update module version to 1.1.6 and update/installation will now clean the module folder of old files.

+++ Install-StratoHiDriveUtils.ps1
@@ -121,7 +121,12 @@ try {
 
 	Get-Module -Name $moduleName | Remove-Module -Force -ErrorAction SilentlyContinue
 	foreach ($candidate in $moduleCandidates) {
-		Remove-Item -LiteralPath $candidate.Path -Recurse -Force -ErrorAction Stop
+		if (Test-Path -LiteralPath $candidate.Path) {
+			Remove-Item -LiteralPath $candidate.Path -Recurse -Force -ErrorAction Stop
+		}
+	}
+	if (Test-Path -LiteralPath $targetPath) {
+		Remove-Item -LiteralPath $targetPath -Recurse -Force -ErrorAction Stop
 	}
 	New-Item -ItemType Directory -Path $targetPath -Force -ErrorAction Stop | Out-Null
